05_nettoyage_agregation_bdd/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "preparation-bdd"
version = "0.1.0"
description = "Exploration rapide de fichiers JSON/CSV"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "orjson>=3.10",
  "ijson>=3.3",
  "pandas>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "ipykernel>=6.29",
  "pyarrow>=15",
  "rapidfuzz>=3.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
